Extract tests for hasPrototype to own file
